refactor: Rename `…kage/compile/events/websockets/lib/routeResponses.js`
