+++ lib/plugins/aws/package/compile/events/websockets/index.js
@@ -5,7 +5,7 @@ const BbPromise = require('bluebird');
 const validate = require('./lib/validate');
 const compileApi = require('./lib/api');
 const compileIntegrations = require('./lib/integrations');
-const compileRouteResponses = require('./lib/routeResponses');
+const compileRouteResponses = require('./lib/route-responses');
 const compilePermissions = require('./lib/permissions');
 const compileRoutes = require('./lib/routes');
 const compileDeployment = require('./lib/deployment');
