+++ test/unit/lib/plugins/aws/package/compile/events/websockets/lib/route-responses.test.js
